Refactor: 선수 추가 버프 계산 함수 분리
